feat: collapse mwc slatepack service into mwc wallet
needs more testing
